+++ walkeeper/src/lib.rs
@@ -4,6 +4,7 @@ use std::path::PathBuf;
 use std::time::Duration;
 
 pub mod pq_protocol;
+pub mod receive_wal;
 pub mod replication;
 pub mod s3_offload;
 pub mod send_wal;
